Patches from Ross Simpson.  From his email:
"First off, I wanted to have the concept of a final tear-down method -- a method called when all tests are done running, just before the results are returned.  I had a little trouble deciding where this should be implemented; I chose the _done() method in jsUnitTestManager.js.  If "top.suiteTearDown" is defined as a method, it will be called at that point.


Second, I needed a way to get some information (generated at test time) into the results, for use later.  I chose to add a property named "userProperty" to the BrowserResult classes.  The user is responsible for setting a top-level variable ( top.userProperty); main-results.html sets the property in the results form submitted back to the acceptor.

I updated the BrowserResultTestCase to reflect the change.

If you're interested, the reasoning behind this is to support javascript code coverage statistics gathering.  I instrument my javascript before the tests are run; statistics are gathered during test runs, and at completion (in my top.suiteTearDown() method), the coverage data is serialized and sent back as the userProperty string.  Further changes I'll submit to you add the userProperty string to the server-side results.  I've made all my changes to your code generic, so they can be used for other purposes."